requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.3
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
schedule==0.6.0
six==1.15.0
Werkzeug==1.0.1
